+++ convert_rule.go
@@ -2,8 +2,6 @@ package ojosama
 
 import (
 	"regexp"
-
-	"github.com/ikawaha/kagome/v2/tokenizer"
 )
 
 // convertRule は 単独のTokenに対して、Conditionsがすべてマッチしたときに変換するルール。
@@ -19,30 +17,73 @@ type convertRule struct {
 	Value                        string
 }
 
+/*
+英語 文法 品詞
+https://ja.wikibooks.org/wiki/%E8%8B%B1%E8%AA%9E/%E6%96%87%E6%B3%95/%E5%93%81%E8%A9%9E
+
+品詞 part of speech (pos)
+*/
+var (
+	posPronounGeneral            = []string{"名詞", "代名詞", "一般"}
+	posNounsGeneral              = []string{"名詞", "一般"}
+	posAdnominalAdjective        = []string{"連体詞"}
+	posAdjectivesSelfSupporting  = []string{"形容詞", "自立"}
+	posInterjection              = []string{"感動詞"}
+	posVerbIndependence          = []string{"動詞", "自立"}
+	posSentenceEndingParticle    = []string{"助詞", "終助詞"}
+	posSubPostpositionalParticle = []string{"助詞", "副助詞"}
+	posAssistantParallelParticle = []string{"助詞", "並立助詞"}
+	posAuxiliaryVerb             = []string{"助動詞"}
+	posNounsSaDynamic            = []string{"名詞", "サ変接続"}
+)
+
+func newRule(features []string, surface, value string) convertRule {
+	return convertRule{
+		Conditions: convertConditions{
+			{
+				Type:  convertTypeFeatures,
+				Value: features,
+			},
+			{
+				Type:  convertTypeSurface,
+				Value: []string{surface},
+			},
+		},
+		Value: value,
+	}
+}
+
+func newRulePronounGeneral(surface, value string) convertRule {
+	return newRule(posPronounGeneral, surface, value)
+}
+
+func newRuleNounsGeneral(surface, value string) convertRule {
+	return newRule(posNounsGeneral, surface, value)
+}
+
+func newRuleAdnominalAdjective(surface, value string) convertRule {
+	return newRule(posAdnominalAdjective, surface, value)
+}
+
+func newRuleAdjectivesSelfSupporting(surface, value string) convertRule {
+	return newRule(posAdjectivesSelfSupporting, surface, value)
+}
+
+func newRuleInterjection(surface, value string) convertRule {
+	return newRule(posInterjection, surface, value)
+}
+
+func (c convertRule) disablePrefix(v bool) convertRule {
+	c.DisablePrefix = v
+	return c
+}
+
 // continuousConditionsConvertRule は連続する条件がすべてマッチしたときに変換するルール。
 type continuousConditionsConvertRule struct {
 	Conditions     []convertConditions
 	AppendLongNote bool
 	Value          string
 }
-
-// convertConditions は変換条件のスライス。
-//
-// この型の評価をする側は、すべての条件が true の時に true として解釈する。
-// 一つでも false が混じったら false として解釈する。
-type convertConditions []convertCondition
-
-type convertType int
-
-// convertCondition は変換に使う条件。
-type convertCondition struct {
-	Type        convertType
-	Value       []string
-	ValueRegexp *regexp.Regexp // オプション。設定されてる時だけ使う
-}
-
-// meaningType は言葉の意味分類。
-type meaningType int
 
 // sentenceEndingParticleConvertRule は「名詞」＋「動詞」＋「終助詞」の組み合わせによる変換ルール。
 type sentenceEndingParticleConvertRule struct {
@@ -52,6 +93,9 @@ type sentenceEndingParticleConvertRule struct {
 	sentenceEndingParticle map[meaningType][]convertConditions // 終助詞
 	value                  map[meaningType][]string
 }
+
+// meaningType は言葉の意味分類。
+type meaningType int
 
 const (
 	convertTypeSurface convertType = iota + 1
@@ -674,208 +718,3 @@ var (
 		newRuleInterjection("ふふふ", "ほほほ"),
 	}
 )
-
-func (c *convertCondition) equalsTokenData(data tokenizer.TokenData) bool {
-	switch c.Type {
-	case convertTypeFeatures:
-		if equalsFeatures(data.Features, c.Value) {
-			return true
-		}
-	case convertTypeSurface:
-		if c.ValueRegexp != nil {
-			return c.ValueRegexp.MatchString(data.Surface)
-		}
-		if data.Surface == c.Value[0] {
-			return true
-		}
-	case convertTypeBaseForm:
-		if c.ValueRegexp != nil {
-			return c.ValueRegexp.MatchString(data.BaseForm)
-		}
-		if data.BaseForm == c.Value[0] {
-			return true
-		}
-	}
-	return false
-}
-
-func (c *convertCondition) notEqualsTokenData(data tokenizer.TokenData) bool {
-	switch c.Type {
-	case convertTypeFeatures:
-		if !equalsFeatures(data.Features, c.Value) {
-			return true
-		}
-	case convertTypeSurface:
-		if c.ValueRegexp != nil {
-			return !c.ValueRegexp.MatchString(data.Surface)
-		}
-		if data.Surface != c.Value[0] {
-			return true
-		}
-	case convertTypeBaseForm:
-		if c.ValueRegexp != nil {
-			return !c.ValueRegexp.MatchString(data.BaseForm)
-		}
-		if data.BaseForm != c.Value[0] {
-			return true
-		}
-	}
-	return false
-}
-
-// matchAllTokenData は data がすべての c と一致した時に true を返す。
-func (c *convertConditions) matchAllTokenData(data tokenizer.TokenData) bool {
-	for _, cond := range *c {
-		if cond.notEqualsTokenData(data) {
-			return false
-		}
-	}
-	return true
-}
-
-// matchAnyTokenData は data がいずれかの c と一致した時に true を返す。
-func (c *convertConditions) matchAnyTokenData(data tokenizer.TokenData) bool {
-	for _, cond := range *c {
-		if cond.equalsTokenData(data) {
-			return true
-		}
-	}
-	return false
-}
-
-func matchAnyMultiConvertConditions(ccs []convertConditions, data tokenizer.TokenData) bool {
-	for _, cs := range ccs {
-		// 1つでも一致すればOK
-		if cs.matchAllTokenData(data) {
-			return true
-		}
-	}
-	return false
-}
-
-/*
-英語 文法 品詞
-https://ja.wikibooks.org/wiki/%E8%8B%B1%E8%AA%9E/%E6%96%87%E6%B3%95/%E5%93%81%E8%A9%9E
-
-品詞 part of speech (pos)
-*/
-var (
-	posPronounGeneral            = []string{"名詞", "代名詞", "一般"}
-	posNounsGeneral              = []string{"名詞", "一般"}
-	posAdnominalAdjective        = []string{"連体詞"}
-	posAdjectivesSelfSupporting  = []string{"形容詞", "自立"}
-	posInterjection              = []string{"感動詞"}
-	posVerbIndependence          = []string{"動詞", "自立"}
-	posSentenceEndingParticle    = []string{"助詞", "終助詞"}
-	posSubPostpositionalParticle = []string{"助詞", "副助詞"}
-	posAssistantParallelParticle = []string{"助詞", "並立助詞"}
-	posAuxiliaryVerb             = []string{"助動詞"}
-	posNounsSaDynamic            = []string{"名詞", "サ変接続"}
-)
-
-func newCond(features []string, surface string) convertConditions {
-	return convertConditions{
-		{
-			Type:  convertTypeFeatures,
-			Value: features,
-		},
-		{
-			Type:  convertTypeSurface,
-			Value: []string{surface},
-		},
-	}
-}
-
-func newCondRe(features []string, re *regexp.Regexp) convertConditions {
-	return convertConditions{
-		{
-			Type:  convertTypeFeatures,
-			Value: features,
-		},
-		{
-			Type:        convertTypeSurface,
-			ValueRegexp: re,
-		},
-	}
-}
-
-func newCondSentenceEndingParticle(surface string) convertConditions {
-	return convertConditions{
-		{
-			Type:  convertTypeFeatures,
-			Value: posSentenceEndingParticle,
-		},
-		{
-			Type:  convertTypeSurface,
-			Value: []string{surface},
-		},
-	}
-}
-
-func newCondAuxiliaryVerb(surface string) convertConditions {
-	return convertConditions{
-		{
-			Type:  convertTypeFeatures,
-			Value: posAuxiliaryVerb,
-		},
-		{
-			Type:  convertTypeSurface,
-			Value: []string{surface},
-		},
-	}
-}
-
-func newConds(surfaces []string) []convertConditions {
-	var c []convertConditions
-	for _, s := range surfaces {
-		cc := convertConditions{
-			{
-				Type:  convertTypeSurface,
-				Value: []string{s},
-			},
-		}
-		c = append(c, cc)
-	}
-	return c
-}
-
-func newRule(features []string, surface, value string) convertRule {
-	return convertRule{
-		Conditions: convertConditions{
-			{
-				Type:  convertTypeFeatures,
-				Value: features,
-			},
-			{
-				Type:  convertTypeSurface,
-				Value: []string{surface},
-			},
-		},
-		Value: value,
-	}
-}
-
-func newRulePronounGeneral(surface, value string) convertRule {
-	return newRule(posPronounGeneral, surface, value)
-}
-
-func newRuleNounsGeneral(surface, value string) convertRule {
-	return newRule(posNounsGeneral, surface, value)
-}
-
-func newRuleAdnominalAdjective(surface, value string) convertRule {
-	return newRule(posAdnominalAdjective, surface, value)
-}
-
-func newRuleAdjectivesSelfSupporting(surface, value string) convertRule {
-	return newRule(posAdjectivesSelfSupporting, surface, value)
-}
-
-func newRuleInterjection(surface, value string) convertRule {
-	return newRule(posInterjection, surface, value)
-}
-
-func (c convertRule) disablePrefix(v bool) convertRule {
-	c.DisablePrefix = v
-	return c
-}
